more work on multi-currency on the backend side
