+++ lib/banchan/studios/studio.ex
@@ -5,96 +5,8 @@ defmodule Banchan.Studios.Studio do
   import Ecto.Changeset
 
   alias Banchan.Identities
-
+  alias Banchan.Studios.Common
   alias Banchan.Uploads.Upload
-
-  # https://stripe.com/docs/connect/cross-border-payouts#supported-countries
-  # But also the US.
-  @supported_countries [
-    "Antigua and Barbuda": :AG,
-    Argentina: :AR,
-    Australia: :AU,
-    Austria: :AT,
-    Bahrain: :BH,
-    Bangladesh: :BD,
-    Belgium: :BE,
-    Benin: :BJ,
-    Bolivia: :BO,
-    Bulgaria: :BG,
-    Canada: :CA,
-    Chile: :CL,
-    Colombia: :CO,
-    "Costa Rica": :CR,
-    "Côte d'Ivoire": :CI,
-    Croatia: :HR,
-    Cyprus: :CY,
-    "Czech Republic": :CZ,
-    Denmark: :DK,
-    "Dominican Republic": :DO,
-    Egypt: :EG,
-    Estonia: :EE,
-    Finland: :FI,
-    France: :FR,
-    Gambia: :GM,
-    Germany: :DE,
-    Ghana: :GH,
-    Greece: :GR,
-    Guatemala: :GT,
-    Guyana: :GY,
-    "Hong Kong": :HK,
-    Hungary: :HU,
-    Iceland: :IS,
-    India: :IN,
-    Indonesia: :ID,
-    Ireland: :IE,
-    Israel: :IL,
-    Italy: :IT,
-    Jamaica: :JM,
-    Japan: :JP,
-    Kenya: :KE,
-    Kuwait: :KW,
-    Latvia: :LV,
-    Liechtenstein: :LI,
-    Lithuania: :LT,
-    Luxembourg: :LU,
-    Malta: :MT,
-    Mauritius: :MU,
-    Mexico: :MX,
-    Monaco: :MC,
-    Morocco: :MA,
-    Namibia: :NA,
-    Netherlands: :NL,
-    "New Zealand": :NZ,
-    Niger: :NE,
-    Norway: :NO,
-    Paraguay: :PY,
-    Peru: :PE,
-    Philippines: :PH,
-    Poland: :PL,
-    Portugal: :PT,
-    Romania: :RO,
-    "San Marino": :SM,
-    "Saudi Arabia": :SA,
-    Senegal: :SN,
-    Serbia: :RS,
-    Singapore: :SG,
-    Slovakia: :SK,
-    Slovenia: :SI,
-    "South Africa": :ZA,
-    "South Korea": :KR,
-    Spain: :ES,
-    "St. Lucia": :LC,
-    Sweden: :SE,
-    Switzerland: :CH,
-    Thailand: :TH,
-    "Trinidad and Tobago": :TT,
-    Tunisia: :TN,
-    Turkey: :TR,
-    "United Arab Emirates": :AE,
-    "United Kingdom": :UK,
-    "United States": :US,
-    Uruguay: :UY
-  ]
 
   schema "studios" do
     field :name, :string
@@ -103,7 +15,9 @@ defmodule Banchan.Studios.Studio do
     field :summary, :string
     field :default_terms, :string
     field :default_template, :string
-    field :country, Ecto.Enum, values: @supported_countries |> Keyword.values()
+    field :country, Ecto.Enum, values: Common.supported_countries() |> Keyword.values()
+    field :default_currency, Ecto.Enum, values: Common.supported_currencies()
+    field :payment_currencies, {:array, Ecto.Enum}, values: Common.supported_currencies()
 
     field :stripe_id, :string
     field :stripe_charges_enabled, :boolean
@@ -125,19 +39,37 @@ defmodule Banchan.Studios.Studio do
   @doc false
   def creation_changeset(studio, attrs) do
     studio
-    |> cast(attrs, [:name, :handle, :description, :country])
-    |> validate_required([:name, :handle, :country])
+    |> cast(attrs, [:name, :handle, :description, :country, :default_currency, :payment_currencies])
+    |> validate_required([:name, :handle, :country, :default_currency, :payment_currencies])
+    |> validate_default_currency(:default_currency, :payment_currencies)
     |> validate_handle_unique(:handle)
   end
 
   @doc false
   def profile_changeset(studio, attrs) do
     studio
-    |> cast(attrs, [:name, :handle, :description, :summary, :default_terms, :default_template])
-    |> validate_required([:name, :handle])
+    |> cast(attrs, [:name, :handle, :description, :summary, :default_currency, :payment_currencies, :default_terms, :default_template])
+    |> validate_required([:name, :handle, :default_currency, :payment_currencies])
     |> validate_markdown(:default_terms)
     |> validate_markdown(:default_template)
+    |> validate_default_currency(:default_currency, :payment_currencies)
     |> validate_handle_unique(:handle)
+  end
+
+  defp validate_default_currency(changeset, default_field, currencies_field)
+  defp validate_default_currency(changeset, default_field, currencies_field) do
+    validate_change(changeset, default_field, fn _, value ->
+    case fetch_field(changeset, currencies_field) do
+      {:changes, currencies} when is_list(currencies) ->
+        if value in currencies do
+          []
+        else
+          [{default_field, "Must be one of the selected payment currencies."}]
+        end
+      _ ->
+        []
+    end
+    end)
   end
 
   defp validate_handle_unique(changeset, field) when is_atom(field) do
@@ -158,9 +90,5 @@ defmodule Banchan.Studios.Studio do
         [{field, "Disallowed HTML detected. Some tags, like <script>, are not allowed."}]
       end
     end)
-  end
-
-  def supported_countries() do
-    @supported_countries
   end
 end
