+++ lib/banchan_web/live/studio_live/new.ex
@@ -10,15 +10,23 @@ defmodule BanchanWeb.StudioLive.New do
   alias Banchan.Studios
   alias Banchan.Studios.Studio
 
-  alias BanchanWeb.Components.Form.{Select, Submit, TextArea, TextInput}
+  alias BanchanWeb.Components.Form.{MultipleSelect, Select, Submit, TextArea, TextInput}
   alias BanchanWeb.Components.Layout
   alias BanchanWeb.Endpoint
 
   @impl true
   def mount(_params, _session, socket) do
+    currencies = Studios.Common.supported_currencies()
+    |> Enum.map(fn currency ->
+      %{name: name, symbol: symbol} = Money.Currency.get(currency)
+      {:"#{name} (#{symbol})", currency}
+    end)
     socket =
       socket
-      |> assign(countries: [{:"Choose your country...", nil} | Studio.supported_countries()])
+      |> assign(
+        countries: [{:"Choose your country...", nil} | Studios.Common.supported_countries()],
+        currencies: [{:"Currencies...", nil} | currencies]
+      )
 
     if is_nil(socket.assigns.current_user.confirmed_at) do
       socket =
@@ -61,6 +69,8 @@ defmodule BanchanWeb.StudioLive.New do
             <TextInput name={:handle} icon="at" opts={required: true} />
             <TextArea name={:description} opts={required: true} />
             <Select name={:country} options={@countries} opts={required: true} />
+            <Select name={:default_currency} options={@currencies} opts={required: true} />
+            <MultipleSelect name={:payment_currencies} options={@currencies} opts={required: true, default_value: :USD} />
             <Submit changeset={@changeset} label="Save" />
           </Form>
         </div>
