+++ priv/repo/seeds.exs
@@ -45,7 +45,9 @@ user = user |> Banchan.Repo.reload()
       handle: "kitteh-studio",
       name: "Kitteh Studio",
       description: "Kitteh-related stuff",
-      country: :US
+      country: "US",
+      default_currency: "USD",
+      payment_currencies: ["USD", "EUR"],
     }
   )
 
@@ -127,7 +129,9 @@ user = user |> Banchan.Repo.reload()
       handle: "kitteh-japan",
       name: "Kitteh Studio in Japan",
       description: "Kitteh-related stuff, Made in Japan™️",
-      country: :JP
+      country: "JP",
+      default_currency: "JPY",
+      payment_currencies: ["JPY", "USD", "KRW"],
     }
   )
